Publish actual version v1.2.3

+++ lib/Service/NodeFolderMappingService.php
@@ -18,11 +18,13 @@ class NodeFolderMappingService
 {
     private $mapper;
     private $fileSystemService;
+    private $nodeTaskService;
 
-    public function __construct(NodeFolderMappingMapper $mapper, FileSystemService $fileSystemService)
+    public function __construct(NodeFolderMappingMapper $mapper, FileSystemService $fileSystemService, NodeTaskMappingService $nodeTaskService)
     {
         $this->mapper = $mapper;
         $this->fileSystemService = $fileSystemService;
+        $this->nodeTaskService = $nodeTaskService;
     }
 
     public function findAll() {
@@ -56,6 +58,14 @@ class NodeFolderMappingService
         }
     }
 
+    public function findMappingByNodeId($nodeId) {
+        try {
+            return $this->getMappingWithFullPath($this->mapper->findMappingForNodeId($nodeId));
+        } catch(Exception $e) {
+            return null;
+        }
+    }
+
     public function find(int $id) {
         try {
             return $this->getMappingWithFullPath($this->mapper->find($id));
@@ -64,21 +74,23 @@ class NodeFolderMappingService
         }
     }
 
-    public function create($ncNodeId, $wrSpaceId) {
+    public function create($ncNodeId, $wrSpaceId, $wrParentId) {
         $nodeFolder = new NodeFolderMapping();
         $nodeFolder->setNcNodeId($ncNodeId);
         $nodeFolder->setWrFolderId($wrSpaceId);
+        $nodeFolder->setWrParentId($wrParentId);
 
         return $this->getMappingWithFullPath($this->mapper->create($nodeFolder));
     }
 
-    public function createForName($ncNodeName, $wrSpaceId) {
+    public function createForName($ncNodeName, $wrSpaceId, $wrParentId) {
         $folder = $this->fileSystemService->getRelativeFsFolderFromSyncRoot($ncNodeName);
 
         if ($folder != null) {
             $nodeFolder = new NodeFolderMapping();
             $nodeFolder->setNcNodeId($folder->getId());
             $nodeFolder->setWrFolderId($wrSpaceId);
+            $nodeFolder->setWrParentId($wrParentId);
 
             return $this->getMappingWithFullPath($this->mapper->create($nodeFolder));
         }
@@ -88,10 +100,42 @@ class NodeFolderMappingService
     public function delete(int $id) {
         try {
             $nodeFolder = $this->mapper->find($id);
-            $this->mapper->delete($nodeFolder);
+
+            //If mapping is found recursively delete all node task mappings below
+            if ($nodeFolder != null) {
+                //Get the start node of the mapping
+                $node = $this->fileSystemService->getFsFolderById($nodeFolder->getNcNodeId());
+                //If the start node still exists then recursively delete the mappings below
+                if ($node != null) {
+                    $subNodes = $this->fileSystemService->getFsSubFoldersOfFsFolder($node);
+                    //Start the recursion of the delete process
+                    foreach ($subNodes as $subNode) {
+                        $this->deleteMappingForSubNode($subNode);
+                    }
+                }
+                //Finally delete the node folder mapping
+                $this->mapper->delete($nodeFolder);
+            }
+
             return $nodeFolder;
         } catch(Exception $e) {
             return null;
+        }
+    }
+
+    private function deleteMappingForSubNode($subNode) {
+        //Try to get the node-task mapping by the ID of the nextcloud node
+        $mapping = $this->nodeTaskService->findMappingByNodeId($subNode->getId());
+        //If any mapping exists delete it
+        if ($mapping != null && $subNode != null) {
+            //Delete the mapping
+            $this->nodeTaskService->delete($mapping);
+            //Get sub nodes of the current node
+            $subNodes = $this->fileSystemService->getFsSubFoldersOfFsFolder($subNode);
+            //Recursively delete all sub nodes mappings
+            foreach ($subNodes as $subNode) {
+                $this->deleteMappingForSubNode($subNode);
+            }
         }
     }
 }
